+++ records/first_SonGyeongI.md
@@ -64,7 +64,7 @@
 - [X] GET /post/myList : 내 글 리스트
   - 내 글 조회는 공개, 비공개 다 볼 수 있어야 한다.
   - [ ] 프론트 - 공개여부 컬럼에서 영어로 true, false가 아닌 한글로 공개, 비공개로 만들기
-  - [ ] /post/myList에서 상세보기할 때 비공개 글도 보이게 하기
+  - [X] /post/myList에서 상세보기할 때 비공개 글도 보이게 하기
 
 글 상세내용 조회
 
@@ -90,9 +90,8 @@
 특정 회원의 글 모아보기
 
 - [X] GET /b/user1 : 회원 user1 의 전체 글 리스트
-  - url : /post/user/user1
   - userPosts.html
-- [ ] GET /b/user1/3 : 회원 user1 의 글 중에서 3번글 상세보기
+- [X] GET /b/user1/3 : 회원 user1 의 글 중에서 3번글 상세보기
   - 글 상세보기로 들어가지는데 왜 필요한지 생각해보기
 
 > 폼
@@ -118,5 +117,5 @@
 - 세세하게 예외 처리, 프론트, 깔끔한 코드 등 많이 신경 못 썼습니다.
 
 [특이사항]
-- 내 글 목록 조회 경우 로그인 후에 nav방 우측 상단에 'welcome user1'을 누르면 나타난다.
+- 내 글 목록 조회 경우 로그인 후에 nav 우측 상단에 'welcome user1'을 누르면 나타난다.
 - 특정 회원의 글 모아보기 경우 게시글 목록에서 작성자 이름을 누르면 나타난다.
