+++ src/main/java/com/ll/medium/domain/post/post/repository/PostRepository.java
@@ -15,6 +15,7 @@ public interface PostRepository extends JpaRepository<Post, Long> {
     // 특정 회원(username)의 전체 글 리스트
     Page<Post> findByIsPublishedTrueAndAuthorId(Long authorId, Pageable pageable);
 
+
     // 게시글 목록, isPublished가 TRUE인 공개된 글만 가져오기
     List<Post> findAllByIsPublishedTrue();
