+++ src/main/java/com/ll/medium/domain/post/post/service/PostService.java
@@ -86,19 +86,20 @@ public class PostService {
 
     // 게시글 상세 조회
     public Post getDetailPost(Long id) {
-        // 작성자와 로그인한 author가 같다면 비공개글 보여주기
-        Post post = postRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("해당 하는 게시글이 없습니다.")
-        );
+        Post post = getPost(id);
 
-        if (rq.getMember().getUsername().equals(post.getAuthor().getUsername())) {
-            return post;
+        // 로그인 하지 않은 유저, 로그인 했지만 작성자가 아닌 유저 isPublished True만 줘야함
+        // 하지만 이미 리스트에는 True만 있음/ url로 직접 작성해서 들어온 사람 막기
+        if (rq.getMember() == null
+        || !(post.getAuthor().getUsername().equals(rq.getMember().getUsername()))) {
+            // 공개된 글 보여주기
+            return postRepository.findByIsPublishedTrueAndId(id).orElseThrow(
+                    () -> new IllegalArgumentException("해당 하는 게시글이 없습니다.")
+            );
         }
 
-        // 공개된 글 보여주기
-        return postRepository.findByIsPublishedTrueAndId(id).orElseThrow(
-                () -> new IllegalArgumentException("해당 하는 게시글이 없습니다.")
-        );
+        // 작성자와 로그인한 author가 같다면 비공개글 보여주기
+        return post;
     }
 
     // 게시글 수정
@@ -111,6 +112,7 @@ public class PostService {
     }
 
     // 게시글 삭제
+    @Transactional
     public void deletePost(Post post) {
         postRepository.delete(post);
     }
