+++ src/main/resources/templates/domain/post/post/list.html
@@ -32,7 +32,7 @@
                 </span>-->
             </td>
             <td>
-                <a th:if="${post.author != null}" th:href="@{|/post/user/${post.author.username}|}" th:text="${post.author.username}"></a>
+                <a th:if="${post.author != null}" th:href="@{|/b/${post.author.username}|}" th:text="${post.author.username}"></a>
             </td>
             <!--            날짜를 보기 좋게 출력하기 위해 타임리프의 #temporals.format 유틸리티를 사용-->
             <td th:text="${#temporals.format(post.createdDate, 'yyyy-MM-dd HH:mm')}"></td>
